fix: Add logging for stela settings and update troubleshooting docs
This commit adds logging in the configuration init file to log the stela settings being used and updates the troubleshooting docs to include information on how to handle Errors related to Stela not finding values for variables in the .env file. This was done to provide users with better troubleshooting abilities and improve user experience.